+++ unionfind/src/test/java/algorithm/UnionFindTest.java
@@ -5,7 +5,7 @@ import java.util.HashSet;
 
 import org.junit.jupiter.api.Test;
 
-import actors.Event;
+import entities.Event;
 import regular_union_find.UnionFind;
 
 class UnionFindTest {
